requirements.txt
numpy
pandas
gensim
spacy
scipy
matplotlib
seaborn
umap-learn
hdbscan
tqdm
zstandard
orjson
argparse